Update CI to build using MSVC 2019
GitHub are going to be removing their windows-2016 images in March 2022,
and Qt 6 only supports MSVC 2019 or mingw81 on Windows, so prebuilt Qt 6
binaries are only available for them.

Binary compatibility between MSVC 2017 and 2019 is such that you can
link binaries compiled with both, so long as the linking is done with
the more recent toolset: this means that we can use Qt 6's MSVC 2019
binaries with our libloot MSVC 2017 binary so long as we run MSVC 2019
for LOOT's CI builds. The details are here:
<https://docs.microsoft.com/en-us/cpp/porting/binary-compat-2015-2017?view=msvc-170>.

I've been using MSVC 2019 locally for a while (at least most of this
year) without any problems. The MSVC redistributable link doesn't need
to be changed because it already points to the 2019-compatible version.